- Create Header - Create Footer - Create Details breed page - Create landing page - Create Api Documentation page - Add react-router-dom
