+++ src/ApiDocumentation.jsx
@@ -1,0 +1,5 @@
+import React from 'react';
+
+export default function ApiDocumentation() {
+  return <h2>Api Documentation</h2>;
+}
