+++ src/DetailsBreed.jsx
@@ -1,0 +1,5 @@
+import React from 'react';
+
+export default function DetailsBreed() {
+  return <h2>Breed Detail</h2>;
+}
