+++ src/Header.css
@@ -1,0 +1,7 @@
+.vertical-centered-li {
+  padding-top: 0.5em;
+}
+
+.vertical-centered-li a {
+  padding-right: 1em;
+}
