+++ src/Spinner.jsx
@@ -1,7 +1,7 @@
 export default function Spinner() {
   return (
-    <div class='spinner-border' role='status'>
-      <span class='visually-hidden'>Loading...</span>
+    <div className='spinner-border' role='status'>
+      <span className='visually-hidden'>Loading...</span>
     </div>
   );
 }
