+++ src/Welcome.css
@@ -1,0 +1,23 @@
+.Welcome {
+  text-align: center;
+  background-color: #282c34;
+  min-height: 100vh;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  font-size: calc(10px + 2vmin);
+  color: white;
+}
+
+.Welcome-img {
+  height: 300px;
+  pointer-events: none;
+  border-radius: 50%;
+}
+
+@media (prefers-reduced-motion: no-preference) {
+  .Welcome-img {
+    animation: App-logo-spin infinite 20s linear;
+  }
+}
