+++ src/Welcome.jsx
@@ -1,0 +1,28 @@
+import React from 'react';
+import './Welcome.css';
+import Spinner from './Spinner';
+import useFetch from './services/useFetch';
+import { Link } from 'react-router-dom';
+
+export default function Welcome() {
+  const { data: cat, error } = useFetch('images/search');
+
+  if (error) throw error;
+  return (
+    <div className='Welcome'>
+      <h1>Welcome to Cat's catalog breeds</h1>
+      <p>This is the Cat of the day...</p>
+      {cat && cat.length > 0 ? (
+        <img src={cat[0].url} className='Welcome-img' alt={cat[0].id} />
+      ) : (
+        <span>
+          <h3> Wait for it...</h3>
+          <Spinner />
+        </span>
+      )}
+      <Link to='/App/' className='btn btn-primary mt-3'>
+        View all cat breeds
+      </Link>
+    </div>
+  );
+}
